# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * World App API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserVerificationLevel =
  (typeof UserVerificationLevel)[keyof typeof UserVerificationLevel];

export const UserVerificationLevel = {
  none: "none",
  device: "device",
  orb: "orb",
} as const;

export interface User {
  id: string;
  username: string;
  displayName: string;
  avatarUrl?: string | null;
  worldId?: string | null;
  verificationLevel: UserVerificationLevel;
  walletAddress: string;
  createdAt: string;
  isVerified: boolean;
  country?: string | null;
  bio?: string | null;
}

export interface UpdateUserBody {
  displayName?: string;
  bio?: string;
  country?: string;
}

export interface Wallet {
  address: string;
  totalBalanceUsd: number;
  wldBalance: string;
  usdcBalance: string;
  ethBalance: string;
  pendingGrants: number;
  nextGrantDate?: string | null;
}

export interface Token {
  symbol: string;
  name: string;
  balance: string;
  balanceUsd: number;
  price: number;
  priceChange24h: number;
  iconUrl?: string | null;
  contractAddress?: string | null;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  send: "send",
  receive: "receive",
  grant: "grant",
  swap: "swap",
} as const;

export type TransactionStatus =
  (typeof TransactionStatus)[keyof typeof TransactionStatus];

export const TransactionStatus = {
  pending: "pending",
  completed: "completed",
  failed: "failed",
} as const;

export interface Transaction {
  id: string;
  type: TransactionType;
  status: TransactionStatus;
  amount: string;
  amountUsd: number;
  tokenSymbol: string;
  fromAddress?: string | null;
  toAddress?: string | null;
  fromUsername?: string | null;
  toUsername?: string | null;
  txHash?: string | null;
  note?: string | null;
  createdAt: string;
}

export interface CreateTransactionBody {
  toAddress?: string;
  toUsername?: string;
  amount: string;
  tokenSymbol: string;
  note?: string;
}

export type VerificationStatusLevel =
  (typeof VerificationStatusLevel)[keyof typeof VerificationStatusLevel];

export const VerificationStatusLevel = {
  none: "none",
  device: "device",
  orb: "orb",
} as const;

export interface VerificationStatus {
  level: VerificationStatusLevel;
  isVerified: boolean;
  verifiedAt?: string | null;
  nullifierHash?: string | null;
  expiresAt?: string | null;
  canClaimGrant: boolean;
}

export type VerificationSessionStatus =
  (typeof VerificationSessionStatus)[keyof typeof VerificationSessionStatus];

export const VerificationSessionStatus = {
  pending: "pending",
  scanning: "scanning",
  processing: "processing",
  complete: "complete",
  failed: "failed",
} as const;

export interface VerificationSession {
  sessionId: string;
  qrCode: string;
  expiresAt: string;
  status: VerificationSessionStatus;
}

export interface CompleteVerificationBody {
  sessionId: string;
  proof?: string;
  nullifierHash?: string;
}

export type AppCategory = (typeof AppCategory)[keyof typeof AppCategory];

export const AppCategory = {
  defi: "defi",
  social: "social",
  gaming: "gaming",
  nft: "nft",
  utility: "utility",
} as const;

export interface App {
  id: string;
  name: string;
  description: string;
  iconUrl?: string | null;
  category: AppCategory;
  url: string;
  isVerified: boolean;
  isFeatured: boolean;
  userCount: number;
  rating?: number | null;
  requiresWorldId: boolean;
  tags: string[];
}

export type FeaturedAppsCategoriesItem = {
  name: string;
  slug: string;
  count: number;
};

export interface FeaturedApps {
  featured: App[];
  categories: FeaturedAppsCategoriesItem[];
  trending: App[];
}

export type ContactVerificationLevel =
  (typeof ContactVerificationLevel)[keyof typeof ContactVerificationLevel];

export const ContactVerificationLevel = {
  none: "none",
  device: "device",
  orb: "orb",
} as const;

export interface Contact {
  id: string;
  username: string;
  displayName: string;
  avatarUrl?: string | null;
  walletAddress: string;
  isVerified: boolean;
  verificationLevel: ContactVerificationLevel;
  lastTransactionAt?: string | null;
}

export interface CreateContactBody {
  username: string;
  walletAddress?: string;
}

export type NotificationType =
  (typeof NotificationType)[keyof typeof NotificationType];

export const NotificationType = {
  transaction: "transaction",
  grant: "grant",
  verification: "verification",
  app: "app",
  system: "system",
} as const;

export type NotificationData = { [key: string]: unknown } | null;

export interface Notification {
  id: string;
  type: NotificationType;
  title: string;
  body: string;
  isRead: boolean;
  data?: NotificationData;
  createdAt: string;
}

export interface DashboardStats {
  totalValueUsd: number;
  wldBalance: string;
  verifiedUsers: number;
  grantsClaimed: number;
  recentTransactionCount: number;
  unreadNotifications: number;
  wldPriceUsd: number;
  wldPriceChange24h: number;
}

export type PriceChartDataPointsItem = {
  timestamp: string;
  price: number;
};

export interface PriceChart {
  symbol: string;
  period: string;
  currentPrice: number;
  priceChange: number;
  priceChangePercent: number;
  dataPoints: PriceChartDataPointsItem[];
}

export type ListTransactionsParams = {
  limit?: number;
  offset?: number;
  type?: ListTransactionsType;
};

export type ListTransactionsType =
  (typeof ListTransactionsType)[keyof typeof ListTransactionsType];

export const ListTransactionsType = {
  send: "send",
  receive: "receive",
  grant: "grant",
  swap: "swap",
} as const;

export type ListAppsParams = {
  category?: ListAppsCategory;
  search?: string;
};

export type ListAppsCategory =
  (typeof ListAppsCategory)[keyof typeof ListAppsCategory];

export const ListAppsCategory = {
  all: "all",
  defi: "defi",
  social: "social",
  gaming: "gaming",
  nft: "nft",
  utility: "utility",
} as const;

export type GetTokenPricesParams = {
  symbol: string;
  period?: GetTokenPricesPeriod;
};

export type GetTokenPricesPeriod =
  (typeof GetTokenPricesPeriod)[keyof typeof GetTokenPricesPeriod];

export const GetTokenPricesPeriod = {
  "1d": "1d",
  "1w": "1w",
  "1m": "1m",
  "3m": "3m",
  "1y": "1y",
} as const;