# lib/api-zod/src/generated/types/app.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * World App API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AppCategory } from "./appCategory";

export interface App {
  id: string;
  name: string;
  description: string;
  iconUrl?: string | null;
  category: AppCategory;
  url: string;
  isVerified: boolean;
  isFeatured: boolean;
  userCount: number;
  rating?: number | null;
  requiresWorldId: boolean;
  tags: string[];
}