# lib/api-zod/src/generated/types/createContactBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * World App API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateContactBody {
  username: string;
  walletAddress?: string;
}