# lib/api-zod/src/generated/types/priceChart.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * World App API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PriceChartDataPointsItem } from "./priceChartDataPointsItem";

export interface PriceChart {
  symbol: string;
  period: string;
  currentPrice: number;
  priceChange: number;
  priceChangePercent: number;
  dataPoints: PriceChartDataPointsItem[];
}