# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * World App API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserVerificationLevel } from "./userVerificationLevel";

export interface User {
  id: string;
  username: string;
  displayName: string;
  avatarUrl?: string | null;
  worldId?: string | null;
  verificationLevel: UserVerificationLevel;
  walletAddress: string;
  createdAt: Date;
  isVerified: boolean;
  country?: string | null;
  bio?: string | null;
}