# lib/api-zod/src/generated/types/verificationStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * World App API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VerificationStatusLevel } from "./verificationStatusLevel";

export interface VerificationStatus {
  level: VerificationStatusLevel;
  isVerified: boolean;
  verifiedAt?: Date | null;
  nullifierHash?: string | null;
  expiresAt?: Date | null;
  canClaimGrant: boolean;
}